US119012: refactor: move user table logic out of Data
